+++ src/Settings/Settings.php
@@ -16,7 +16,7 @@ class Settings
     private const CONFIG_FILE_NAME = 'cli-config.json';
 
     public const TRANSIP_API_ENDPOINT = 'https://api.transip.nl/v6';
-    public const TRANSIP_CLI_VERSION = '6.19.0';
+    public const TRANSIP_CLI_VERSION = '6.22.0';
 
     /**
      * @var string