+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 CHANGELOG
 =========
+
+6.22.0
+-----
+* Added domain handover
+
+6.21.0
+-----
+* Added default domain contacts endpoint
+
 6.19.0
 -----
 * Added includes to domain:getbyname and domain:getall